Add support for Ohai7 compatible plugin
The plugin format changed between Ohai6 and Ohai7 causing some upgrade problems for people using Vagrant.

This changeset addresses the issue by having an ohai7 compatible plugin in addition to the original ohai6 plugin. It checks the Ohai major version to decide which plugin file to use.

I had to replace the single remote directory resource with three resources. One for the directory, one for the README, and one for the plugin file.